Refactored most of the frane and panes to the Frame.cs class
