+++ oopProto/Entities/Services/RoomService.cs
@@ -87,12 +87,12 @@ public class RoomService
         if (this.currentRoom.EastId != 0) east = true;
         if (this.currentRoom.WestId != 0) west = true;
         
-        currentRoomAvailablePath = $"Available Paths:                           \n" +
-                                   $"                  \u2191                   \n" +
-                                   $"             north-{north}                 \n" +
-                                   $"\u2190 west-{west}------+------{east}-east \u2192\n" +
-                                   $"             south-{south}                 \n" +
-                                   $"                  \u2193                   \n";
+        currentRoomAvailablePath = $"-----------------------------------------\n" +
+                                   $"|                   \u2191                   |\n" +
+                                   $"|              north-{north}               |\n" +
+                                   $"| \u2190 west-{west}------+------{east}-east \u2192 |\n" +
+                                   $"|              south-{south}              |\n" +
+                                   $"|                   \u2193                   |\n";
         
         
         return currentRoomAvailablePath;
